web-animations/animatable/animatable-get-current-animations.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Animatable.getCurentAnimations() method tests</title>
<meta name="assert" content="Returns the set of current Animation objects that have an animation effect whose animation target is this object">
<link rel="help" href="https://web-animations.github.io/web-animations-spec/bikeshed/Overview.html#dom-animatable-getcurrentanimations">
<link rel="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
<body>
<div id="log"></div>
<script>
test(function() {
    var animatable = createDiv(this);
    var player = animatable.animate(null);
    var animations = animatable.getCurrentAnimations();

    assert_array_equals(animations, [player.source], 'Animatable.getCurrentAnimations() ' +
        'method should return all animations whose animation target is this object');
}, 'Test that Animatable.getCurrentAnimations() method returns single animation whose ' +
    'animation target is this object. Animation is created by Animatable.animate() call');

test(function() {
    var animatable = createDiv(this);
    var animation = newAnimation(animatable);
    var animations = animatable.getCurrentAnimations();

    assert_array_equals(animations, [animation], 'Animatable.getCurrentAnimations() ' +
        'method should return all animations whose animation target is this object');
}, 'Test that Animatable.getCurrentAnimations() method returns single animation whose ' +
    'animation target is this object. Animation is created by Animation constructor');

test(function() {
    var animatable = createDiv(this);
    var player1 = animatable.animate(null);
    var player2 = animatable.animate(null);
    var animations = animatable.getCurrentAnimations();

    assert_array_equals(animations, [player2.source, player1.source],
        'Animatable.getCurrentAnimations() method should return all animations whose ' +
        'animation target is this object and sorted by their associated animation effect');
}, 'Test that Animatable.getCurrentAnimations() method returns all animations whose ' +
    'animation target is this object and sorted by their associated animation effect. ' +
    'Animations are created by Animatable.animate() call');

test(function() {
    var animatable = createDiv(this);
    var animation1 = newAnimation(animatable);
    var animation2 = newAnimation(animatable);
    var animations = animatable.getCurrentAnimations();
// FIXME Is this order specified?
    assert_array_equals(animations, [animation2, animation1],
        'Animatable.getCurrentAnimations() method should return all animations whose ' +
        'animation target is this object');
}, 'Test that Animatable.getCurrentAnimations() method returns single animation whose ' +
    'animation target is this object and and sorted by their associated animation effect. ' +
    'Animations are created by Animation constructor');
</script>
</body>
